handler and examples clean up, will be have separate repos
