+++ superagentx/handler/__init__.py
@@ -1,15 +1,6 @@
 from .ai import AIHandler
-from .aws import AWSS3Handler
-from .crm import HubSpotHandler
-from .ecommerce import AmazonHandler, FakeAmazonHandler, FakeFlipkartHandler, FlipkartHandler
 from .elastic_search import ElasticsearchHandler
-from .email import EmailHandler
+from .send_email import EmailHandler
 from .exa_search import ExaHandler
 from .financial_data import FinancialHandler
-from .google import CalenderHandler, GmailHandler
-from .scrape import ScrapeHandler
 from .serper_dev import SerperDevToolHandler
-from .serply_api import SerplyJobHandler
-from .sql import SQLHandler
-from .weather import WeatherHandler
-from .wikipedia import WikipediaHandler
